Fix admin self actions at /me/ route

+++ controllers/userController.js
@@ -79,11 +79,15 @@ exports.search = catchAsync(async (req, res, next) => {
 //  - get user id from params and show all info about user
 exports.getUser = catchAsync(async (req, res, next) => {
   const searchedUser =
-    req.user.role === 'admin'
+    req.user.role === 'admin' && req.route.path !== '/me/'
       ? await User.findById(req.params.id).select(
           '+isHidden +isVerified +isActive'
         )
       : await User.findById(req.user.id);
+
+  if (!searchedUser) {
+    return next(new AppError('No user found', 404));
+  }
 
   res.status(200).json({
     status: 'success',
@@ -101,7 +105,7 @@ exports.getUser = catchAsync(async (req, res, next) => {
 exports.updateUser = catchAsync(async (req, res, next) => {
   // Create error if user POSTs password data
   if (
-    req.user.role !== 'admin' &&
+    req.route.path === '/me/' &&
     (req.body.password || req.body.passwordConfirm)
   ) {
     return next(new AppError('This route is not for password updates.', 400));
@@ -109,7 +113,7 @@ exports.updateUser = catchAsync(async (req, res, next) => {
 
   // Filtered out unwanted fields names that are not allowed to be updated
   const filteredBody =
-    req.user.role !== 'admin' &&
+    req.route.path === '/me/' &&
     filterObj(
       req.body,
       'gender',
@@ -123,7 +127,7 @@ exports.updateUser = catchAsync(async (req, res, next) => {
 
   // Update user document and returned the new one
   const updatedUser =
-    req.user.role === 'admin'
+    req.user.role === 'admin' && req.route.path !== '/me/'
       ? await User.findByIdAndUpdate(req.params.id, req.body, {
           new: true,
           runValidators: true
@@ -132,6 +136,10 @@ exports.updateUser = catchAsync(async (req, res, next) => {
           new: true,
           runValidators: true
         });
+
+  if (!updatedUser) {
+    return next(new AppError('No user found', 404));
+  }
 
   res.status(200).json({
     status: 'success',
@@ -148,7 +156,13 @@ exports.updateUser = catchAsync(async (req, res, next) => {
 //  - delete user from DB
 exports.deleteUser = catchAsync(async (req, res, next) => {
   let user = null;
-  if (req.user.role === 'admin') {
+
+  // Check if admin wants to delete/deactivate himself
+  if (req.user.role === 'admin' && req.route.path === '/me/') {
+    return next(new AppError('Can not delete/deactivate admin', 400));
+  }
+
+  if (req.user.role === 'admin' && req.route.path !== '/me/') {
     user = await User.findByIdAndDelete(req.params.id);
     if (!user) {
       return next(new AppError('No user found with that ID', 404));
